requirements.txt
Flask>=1.1.1
google-auth>=1.5.1
requests>=2.21.0
google-cloud-firestore>=1.3.0

pytest>=5.0.1